style(Spinner): move to components/Spinner.jsx
plus some eslint style fixes
also rename components/Visualizations.js to
  CustomPieChart.jsx
